Add continuous integration (#46)
* Add continuous integration

One workflow, two jobs. The host job runs the full local test estate:
the core C suite, the Unity binaries, both profile transcripts, the
library end-to-end build, the Go CLI tests, and the two TypeScript
library suites. The esp32 job proves the firmware builds against
ESP-IDF v5.5, with the SDK cached between runs; no hardware is
involved. A status badge tops the README.

frothy-repl gains a package-lock.json so its install is reproducible
under npm ci, matching frothy-editor.

* Fix the two failures the first CI run exposed

gcc treats the quoted-text chunk bound as a signed/unsigned comparison
(-Werror=sign-compare); clang locally does not. Compare unsigned.

The firmware recipe sourced the ESP-IDF export script with all output
discarded inside an && chain; under make's default shell on Linux
(dash) the bash-flavored script fails and the chain dies silently.
Force bash for recipes and keep stderr visible when sourcing so a
future failure says why.

* Fix the remaining gcc sign-compare in base_image

Swept every kernel, target, board, and test source with gcc's
-Wsign-compare under both host profiles; this line-buffer bound in
fr_base_compile_source was the only instance left.

* Fix CI cache coverage and the frothy-repl lockfile

The ESP-IDF cache saved only ~/.froth/sdk; the toolchain's Python
virtualenv lives under ~/.espressif, so a cache hit restored an SDK
that could not activate. Cache both paths.

The frothy-repl lockfile was generated with --package-lock-only, which
omits platform-specific optional dependencies, so npm ci desynced on
Linux. Regenerated with a full install.

+++ src/base_image.c
@@ -133,7 +133,7 @@ static fr_err_t fr_base_compile_source(fr_runtime_t *runtime, const char *bytes,
     bool blank = true;
 
     while (i < length && bytes[i] != '\n') {
-      if (n + 1 >= sizeof(line)) {
+      if ((uint32_t)n + 1 >= sizeof(line)) {
         return FR_ERR_CAPACITY;
       }
       if (bytes[i] != ' ' && bytes[i] != '\t' && bytes[i] != '\r') {
